# backend/swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/upload":{"post":{"operationId":"AppController_uploadFile","summary":"Upload an image","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}},"201":{"description":""}},"security":[{"bearer":[]}]}},"/files/{filepath}":{"get":{"operationId":"AppController_seeUploadedFile","summary":"Visualize uploaded file","parameters":[{"name":"filepath","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Login response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["auth"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"200":{"description":"Register response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["auth"]}},"/auth/refresh-token":{"post":{"operationId":"AuthController_refreshToken","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"200":{"description":"Refresh token response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/request-reset-password-email":{"post":{"operationId":"AuthController_requestPasswordReset","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordReqDto"}}}},"responses":{"200":{"description":"Request reset password"}},"tags":["auth"]}},"/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":"Reset password"}},"tags":["auth"]}},"/auth/me":{"get":{"operationId":"AuthController_getAuthenticatedUser","parameters":[],"responses":{"200":{"description":"Get authenticated user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["auth"],"security":[{"bearer":[]}]}}},"info":{"title":"TUNLOG","description":"Tunlog apis","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ResponseDto":{"type":"object","properties":{"data":{"type":"object"}},"required":["data"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"AuthResponseDto":{"type":"object","properties":{"accessToken":{"type":"object"},"refreshToken":{"type":"object"},"user":{"type":"object"}},"required":["accessToken","refreshToken","user"]},"RegisterDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]},"ResetPasswordReqDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"ResetPasswordDto":{"type":"object","properties":{"newPassword":{"type":"string"},"resetPasswordToken":{"type":"string"}},"required":["newPassword","resetPasswordToken"]}}}}